# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/rentals": {
      "post": {
        "operationId": "createRental",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRentalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rental"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Rentals"
        ]
      },
      "get": {
        "operationId": "getRentals",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "roomId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "NOTICE_GIVEN",
                "TERMINATED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Rental"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Rentals"
        ]
      }
    },
    "/api/v1/rentals/{id}": {
      "get": {
        "operationId": "getRental",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rental"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Rentals"
        ]
      },
      "patch": {
        "operationId": "updateRental",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRentalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rental"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Rentals"
        ]
      },
      "delete": {
        "operationId": "deleteRental",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rental"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Rentals"
        ]
      }
    },
...
    "/api/v1/payments": {
      "post": {
        "operationId": "createPayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Payments"
        ]
      },
      "get": {
        "operationId": "getPayments",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "billId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "COMPLETED",
                "FAILED",
                "REFUNDED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Payment"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/{id}": {
      "get": {
        "operationId": "getPayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Payments"
        ]
      }
    },
...
    "/api/v1/notifications": {
      "post": {
        "operationId": "createNotification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Notifications"
        ]
      },
      "get": {
        "operationId": "getNotifications",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "isRead",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "BILL_GENERATED",
                "PAYMENT_REMINDER",
                "MAINTENANCE_UPDATE",
                "CHAT_MESSAGE",
                "ANNOUNCEMENT",
                "SYSTEM"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Notification"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/notifications/unread/count": {
      "get": {
        "operationId": "getUnreadNotificationCount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/notifications/{id}": {
      "get": {
        "operationId": "getNotification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/notifications/{id}/read": {
      "patch": {
        "operationId": "markNotificationAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/notifications/read/all": {
      "patch": {
        "operationId": "markAllNotificationsAsRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Notifications"
        ]
      }
    }
  },
  "info": {
    "title": "TacoHouse API",
    "description": "A comprehensive rental room management system API for managing buildings, rooms, tenants, billing, and payments.\n\n## Response Format\nAll responses follow a standardized format:\n- **status**: HTTP status code\n- **message**: Human-readable message\n- **data**: Response data (object or array)\n- **pagination**: (Optional) Pagination metadata for list endpoints",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3005",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "UserRole": {
        "type": "string",
        "enum": [
          "ADMIN",
          "LANDLORD",
          "TENANT"
        ]
      },
...
      "RentalStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "NOTICE_GIVEN",
          "TERMINATED"
        ]
      },
      "Rental": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "tenantRole": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "roomId": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "noticeDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "monthlyRent": {
            "type": "string",
            "format": "Decimal.js"
          },
          "depositPaid": {
            "type": "string",
            "format": "Decimal.js"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RentalStatus"
              }
            ]
          },
          "contractImages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "tenantId",
          "tenantRole",
          "roomId",
          "startDate",
          "endDate",
          "noticeDate",
          "monthlyRent",
          "depositPaid",
          "status",
          "contractImages",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "PaymentMethod": {
        "type": "string",
        "enum": [
          "CASH",
          "BANK_TRANSFER",
          "STRIPE"
        ]
      },
      "PaymentStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "COMPLETED",
          "FAILED",
          "REFUNDED"
        ]
      },
      "PaymentConfirmation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "paymentId": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "tenantRole": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "tenantConfirmed": {
            "type": "boolean"
          },
          "tenantConfirmedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "landlordConfirmed": {
            "type": "boolean"
          },
          "landlordConfirmedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "proofImages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "paymentId",
          "tenantId",
          "tenantRole",
          "tenantConfirmed",
          "tenantConfirmedAt",
          "landlordConfirmed",
          "landlordConfirmedAt",
          "proofImages",
          "notes",
          "createdAt",
          "updatedAt"
        ]
      },
      "Payment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "billId": {
            "type": "string"
          },
          "amount": {
            "type": "string",
            "format": "Decimal.js"
          },
          "paymentMethod": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentMethod"
              }
            ]
          },
          "paymentDate": {
            "type": "string",
            "format": "date-time"
          },
          "stripePaymentId": {
            "type": "string",
            "nullable": true
          },
          "bankTransferRef": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "receiptImage": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentStatus"
              }
            ]
          },
          "confirmation": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentConfirmation"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "billId",
          "amount",
          "paymentMethod",
          "paymentDate",
          "stripePaymentId",
          "bankTransferRef",
          "notes",
          "receiptImage",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "NotificationType": {
        "type": "string",
        "enum": [
          "BILL_GENERATED",
          "PAYMENT_REMINDER",
          "MAINTENANCE_UPDATE",
          "CHAT_MESSAGE",
          "ANNOUNCEMENT",
          "SYSTEM"
        ]
      },
      "RelatedEntityType": {
        "type": "string",
        "enum": [
          "BILL",
          "PAYMENT",
          "RENTAL",
          "MAINTENANCE_REQUEST",
          "MESSAGE",
          "ROOM",
          "BUILDING"
        ]
      },
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/NotificationType"
              }
            ]
          },
          "isRead": {
            "type": "boolean"
          },
          "readAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "relatedId": {
            "type": "string",
            "nullable": true
          },
          "relatedType": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedEntityType"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "title",
          "message",
          "type",
          "isRead",
          "readAt",
          "relatedId",
          "relatedType",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateRentalDto": {
        "type": "object",
        "properties": {
          "roomId": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "monthlyRent": {
            "type": "number",
            "minimum": 0
          },
          "depositPaid": {
            "type": "number",
            "minimum": 0
          },
          "contractImages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "roomId",
          "tenantId",
          "startDate",
          "monthlyRent"
        ]
      },
      "UpdateRentalDto": {
        "type": "object",
        "properties": {
          "endDate": {
            "type": "string"
          },
          "noticeDate": {
            "type": "string"
          },
          "monthlyRent": {
            "type": "number",
            "minimum": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "NOTICE_GIVEN",
              "TERMINATED"
            ]
          }
        }
      },
...
      "CreatePaymentDto": {
        "type": "object",
        "properties": {
          "billId": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "minimum": 0
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "CASH",
              "BANK_TRANSFER",
              "STRIPE"
            ]
          },
          "transactionId": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "receiptImage": {
            "type": "string"
          }
        },
        "required": [
          "billId",
          "amount",
          "paymentMethod"
        ]
      },
...
      "CreateNotificationDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "BILL_GENERATED",
              "PAYMENT_REMINDER",
              "MAINTENANCE_UPDATE",
              "CHAT_MESSAGE",
              "ANNOUNCEMENT",
              "SYSTEM"
            ]
          },
          "relatedId": {
            "type": "string"
          },
          "relatedType": {
            "type": "string",
            "enum": [
              "BILL",
              "PAYMENT",
              "RENTAL",
              "MAINTENANCE_REQUEST",
              "MESSAGE",
              "ROOM",
              "BUILDING"
            ]
          }
        },
        "required": [
          "userId",
          "title",
          "message",
          "type"
        ]
      }
    }
  }
}